staff-service/requirements.txt
Django==5.1.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
requests==2.31.0
